PHOENIX-274 New masters and new columns

+++ egov/egov-egf/src/main/java/org/egov/web/actions/bill/ContingentBillAction.java
@@ -821,7 +821,7 @@ public class ContingentBillAction extends BaseBillAction {
                         CheckListHelper chkHelper=new CheckListHelper();
                         chkHelper.setName(chk.getAppconfigvalue().getValue());
                         chkHelper.setVal(chk.getChecklistvalue());
-                        chkHelper.setId(chk.getAppconfigvalue().getId());  
+                        chkHelper.setId((chk.getAppconfigvalue().getId()).intValue());  
                         checkListsTable.add(chkHelper);
                 }
                 return cbill;
